CUDA compilation is OK

+++ TPZCudaCalls.cu
@@ -1,5 +1,5 @@
 #include "TPZCudaCalls.h"
-// #include "pzreal.h"
+#include "pzreal.h"
 #include "pzvec.h"
 
 #include "SpectralDecompKernels.h"
@@ -23,8 +23,8 @@ TPZCudaCalls::~TPZCudaCalls() {
 	}
 }
 
-void TPZCudaCalls::Multiply(bool trans, int *m, int *n, int *k, double *A, int *strideA, 
-	double *B, int *strideB,  double *C, int *strideC, double alpha, int nmatrices) {
+void TPZCudaCalls::Multiply(bool trans, int *m, int *n, int *k, REAL *A, int *strideA, 
+	REAL *B, int *strideB,  REAL *C, int *strideC, REAL alpha, int nmatrices) {
 
 	MatrixMultiplicationKernel<<<nmatrices,1>>> (trans, m, n, k, A, strideA, B, strideB, C, strideC, alpha, nmatrices);
 	cudaDeviceSynchronize();
@@ -34,7 +34,7 @@ void TPZCudaCalls::Multiply(bool trans, int *m, int *n, int *k, double *A, int *
 
 }
 
-void TPZCudaCalls::GatherOperation(int n, double *x, double *y, int *id) {
+void TPZCudaCalls::GatherOperation(int n, REAL *x, REAL *y, int *id) {
 	if(cusparse_h == false) {
 		cusparse_h = true;
 		cusparseStatus_t result = cusparseCreate(&handle_cusparse);
@@ -48,7 +48,7 @@ void TPZCudaCalls::GatherOperation(int n, double *x, double *y, int *id) {
 	}	
 }
 
-void TPZCudaCalls::ScatterOperation(int n, double *x, double *y, int *id) {
+void TPZCudaCalls::ScatterOperation(int n, REAL *x, REAL *y, int *id) {
 	if(cusparse_h == false) {
 		cusparse_h = true;
 		cusparseStatus_t result = cusparseCreate(&handle_cusparse);
@@ -76,7 +76,7 @@ void TPZCudaCalls::DaxpyOperation(int n, double alpha, double *x, double *y) {
 	}	
 }
 
-void TPZCudaCalls::SpMV(int opt, int m, int k, int nnz, double alpha, double *csrVal, int *csrRowPtr, int *csrColInd, double *B, double *C) {
+void TPZCudaCalls::SpMV(int opt, int m, int k, int nnz, REAL alpha, REAL *csrVal, int *csrRowPtr, int *csrColInd, REAL *B, REAL *C) {
 	if(cusparse_h == false) {
 		cusparse_h = true;
 		cusparseStatus_t result = cusparseCreate(&handle_cusparse);
@@ -89,7 +89,7 @@ void TPZCudaCalls::SpMV(int opt, int m, int k, int nnz, double alpha, double *cs
 	cusparseSetMatType(descr, CUSPARSE_MATRIX_TYPE_GENERAL);
 	cusparseSetMatIndexBase(descr, CUSPARSE_INDEX_BASE_ZERO);
 
-	double beta = 0.;
+	REAL beta = 0.;
 	cusparseStatus_t result;
 	if(opt == 0) {
 		result = cusparseDcsrmv(handle_cusparse, CUSPARSE_OPERATION_NON_TRANSPOSE, m, k, nnz, &alpha, descr, csrVal, csrRowPtr, csrColInd, B, &beta, C);
@@ -101,9 +101,9 @@ void TPZCudaCalls::SpMV(int opt, int m, int k, int nnz, double alpha, double *cs
 	}	
 }
 
-void TPZCudaCalls::SpMSpM(int opt, int m, int n, int k, int nnzA, double *csrValA, int *csrRowPtrA, int *csrColIndA, 
-														int nnzB, double *csrValB, int *csrRowPtrB, int *csrColIndB, 
-														int nnzC, double *csrValC, int *csrRowPtrC) {
+void TPZCudaCalls::SpMSpM(int opt, int m, int n, int k, int nnzA, REAL *csrValA, int *csrRowPtrA, int *csrColIndA, 
+														int nnzB, REAL *csrValB, int *csrRowPtrB, int *csrColIndB, 
+														int nnzC, REAL *csrValC, int *csrRowPtrC) {
 	if(cusparse_h == false) {
 		cusparse_h = true;
 		cusparseStatus_t result = cusparseCreate(&handle_cusparse);
